requirements.txt
# 核心依赖
python-multipart>=0.0.6
aiofiles>=23.2.1

# 文档处理
python-docx>=1.1.0
openpyxl>=3.1.2
pdfplumber>=0.10.3
PyPDF2>=3.0.1

# 图片处理
Pillow>=10.0.0

# Web 框架 (可选)
fastapi>=0.100.0
uvicorn>=0.23.0

# 日志
loguru>=0.7.0

# 配置管理
pyyaml>=6.0
python-dotenv>=1.0.0
